move metadata wrapper to own file
