+++ cql_tests.py
@@ -17,7 +17,8 @@ from thrift_bindings.v22.ttypes import \
 from thrift_bindings.v22.ttypes import (CfDef, Column, ColumnOrSuperColumn,
                                         Mutation)
 from thrift_tests import get_thrift_client
-from tools import debug, rows_to_list, since, UpdatingClusterMetadataWrapper
+from tools import debug, rows_to_list, since
+from utils.metadata_wrapper import UpdatingClusterMetadataWrapper
 
 
 class CQLTester(Tester):
